Seperate concerns into their own headers
