+++ src/coord.h
@@ -1,0 +1,14 @@
+/* Copyright 2023 Arjun Aravind */
+#ifndef SRC_COORD_H_
+#define SRC_COORD_H_
+
+#include<utility>
+
+namespace sudoku {
+
+typedef std::pair<int, int> Coord;
+int GRID_LEN = 9;
+
+}  // namespace sudoku
+
+#endif  // SRC_COORD_H_
